feat(flight): double-precision FlightState velocity vector for ICBM-range trajectories (#387) (#400)
Upgrades FlightState::vel_body[3] from float[3] to double[3], eliminating
~0.97 m/s accumulated velocity error over a 30-minute burn+coast phase at
Mach 20. Unblocks ballistic force model (#354) and missile guidance (#355).

- Add quatRotateD (file-local, float quat × double vec) for the vel_body →
  pos_world position update; float quatRotate still used for Mach/airspeed
  calculations where aerodynamic precision is sufficient
- kMaxBodySpeed changed to constexpr double to avoid std::clamp deduction
  failure with mixed float/double types
- Float casts added at the two aero boundary sites (wing-sweep Mach and
  relative-airspeed step 5); ground-collision vw array and horizSpd narrowing
  both get explicit float() casts to satisfy -Werror on all three platforms
- WorldBroadcaster::copyFlightStateToTransform casts vel_body to float before
  calling its local quatRotate; MsgEntityEntry::vel and EntityRenderEntry::
  velocity remain float (dead-reckoning over ~16 ms needs no ICBM precision)
- 3 new tests: type precision, integration chain smoke, kMaxBodySpeed clamp

+++ engine/flight/FlightIntegrator.h
@@ -15,7 +15,7 @@ namespace fl {
 // Aircraft state vector — flat/POD for network serialisation.
 struct FlightState {
     double pos_world[3]{};         // world-frame position (m) — double for planet-scale precision
-    float vel_body[3]{};           // body-frame velocity (m/s)
+    double vel_body[3]{};          // body-frame velocity (m/s) — double for ICBM-range precision
     float euler[3]{};              // roll, pitch, yaw (rad) — derived from quaternion
     float omega[3]{};              // body-frame angular rates: p, q, r (rad/s)
     float quat[4]{0, 0, 0, 1};     // body↔world quaternion [x,y,z,w]
